Parameters can be copy-pasted into the classes...
... and public members can again be private.

But then some tables of strings need to be declared inside the classes.
