+++ src/effects/BassTreble.h
@@ -104,7 +104,6 @@ private:
    void UpdateGain(double oldVal, int control );
 
 private:
-public: // TODO remove
    EffectBassTrebleState mMaster;
    std::vector<EffectBassTrebleState> mSlaves;
 
@@ -125,6 +124,15 @@ public: // TODO remove
 
    const EffectParameterMethods& Parameters() const override;
    DECLARE_EVENT_TABLE()
+
+static constexpr EffectParameter Bass{ &EffectBassTreble::mBass,
+   L"Bass",          0.0,     -30.0,   30.0,    1  };
+static constexpr EffectParameter Treble{ &EffectBassTreble::mTreble,
+   L"Treble",        0.0,     -30.0,   30.0,    1  };
+static constexpr EffectParameter Gain{ &EffectBassTreble::mGain,
+   L"Gain",          0.0,     -30.0,   30.0,    1  };
+static constexpr EffectParameter Link{ &EffectBassTreble::mLink,
+   L"Link Sliders",  false,    false,  true,    1  };
 };
 
 #endif
