+++ src/effects/ChangePitch.cpp
@@ -69,12 +69,6 @@ enum {
 
 // Soundtouch is not reasonable below -99% or above 3000%.
 
-namespace {
-static constexpr EffectParameter Percentage{ &EffectChangePitch::m_dPercentChange,
-   L"Percentage", 0.0,  -99.0,   3000.0,  1  };
-static constexpr EffectParameter UseSBSMS{ &EffectChangePitch::mUseSBSMS,
-   L"SBSMS",     false, false,   true,    1  };
-}
 const EffectParameterMethods& EffectChangePitch::Parameters() const
 {
    static CapturedParameters<EffectChangePitch,
