+++ src/effects/ChangeSpeed.cpp
@@ -64,10 +64,6 @@ static const TranslatableStrings kVinylStrings{
 
 // Soundtouch is not reasonable below -99% or above 3000%.
 
-namespace {
-static constexpr EffectParameter Percentage{ &EffectChangeSpeed::m_PercentChange,
-   L"Percentage",    0.0,  -99.0,   4900.0,  1  };
-}
 const EffectParameterMethods& EffectChangeSpeed::Parameters() const
 {
    static CapturedParameters<EffectChangeSpeed,
