+++ src/effects/Compressor.cpp
@@ -56,22 +56,6 @@ enum
    ID_Decay
 };
 
-namespace {
-static constexpr EffectParameter Threshold{ &EffectCompressor::mThreshold,
-   L"Threshold",     -12.0,   -60.0,   -1.0,    1   };
-static constexpr EffectParameter NoiseFloor{ &EffectCompressor::mNoiseFloor,
-   L"NoiseFloor",    -40.0,   -80.0,   -20.0,   0.2   };
-static constexpr EffectParameter Ratio{ &EffectCompressor::mRatio,
-   L"Ratio",         2.0,     1.1,     10.0,    10  };
-static constexpr EffectParameter AttackTime{ &EffectCompressor::mAttackTime,
-   L"AttackTime",    0.2,     0.1,     5.0,     100 };
-static constexpr EffectParameter ReleaseTime{ &EffectCompressor::mDecayTime,
-   L"ReleaseTime",   1.0,     1.0,     30.0,    10  };
-static constexpr EffectParameter Normalize{ &EffectCompressor::mNormalize,
-   L"Normalize",     true,    false,   true,    1   };
-static constexpr EffectParameter UsePeak{ &EffectCompressor::mUsePeak,
-   L"UsePeak",       false,   false,   true,    1   };
-}
 const EffectParameterMethods& EffectCompressor::Parameters() const
 {
    static CapturedParameters<EffectCompressor,
