+++ src/effects/Compressor.h
@@ -74,7 +74,6 @@ private:
    void UpdateUI();
 
 private:
-public: // TODO remove
    double    mRMSSum;
    size_t    mCircleSize;
    size_t    mCirclePos;
@@ -129,6 +128,21 @@ public: // TODO remove
 
    const EffectParameterMethods& Parameters() const override;
    DECLARE_EVENT_TABLE()
+
+static constexpr EffectParameter Threshold{ &EffectCompressor::mThreshold,
+   L"Threshold",     -12.0,   -60.0,   -1.0,    1   };
+static constexpr EffectParameter NoiseFloor{ &EffectCompressor::mNoiseFloor,
+   L"NoiseFloor",    -40.0,   -80.0,   -20.0,   0.2   };
+static constexpr EffectParameter Ratio{ &EffectCompressor::mRatio,
+   L"Ratio",         2.0,     1.1,     10.0,    10  };
+static constexpr EffectParameter AttackTime{ &EffectCompressor::mAttackTime,
+   L"AttackTime",    0.2,     0.1,     5.0,     100 };
+static constexpr EffectParameter ReleaseTime{ &EffectCompressor::mDecayTime,
+   L"ReleaseTime",   1.0,     1.0,     30.0,    10  };
+static constexpr EffectParameter Normalize{ &EffectCompressor::mNormalize,
+   L"Normalize",     true,    false,   true,    1   };
+static constexpr EffectParameter UsePeak{ &EffectCompressor::mUsePeak,
+   L"UsePeak",       false,   false,   true,    1   };
 };
 
 class EffectCompressorPanel final : public wxPanelWrapper
