+++ src/effects/Loudness.cpp
@@ -33,31 +33,12 @@
 
 #include "LoadEffects.h"
 
-enum kNormalizeTargets
-{
-   kLoudness,
-   kRMS,
-   nAlgos
-};
-
-static const EnumValueSymbol kNormalizeTargetStrings[nAlgos] =
+static const EnumValueSymbol kNormalizeTargetStrings[EffectLoudness::nAlgos] =
 {
    { XO("perceived loudness") },
    { XO("RMS") }
 };
 
-namespace {
-static constexpr EffectParameter StereoInd{ &EffectLoudness::mStereoInd,
-   L"StereoIndependent",   false,      false,   true,     1  };
-static constexpr EffectParameter LUFSLevel{ &EffectLoudness::mLUFSLevel,
-   L"LUFSLevel",           -23.0,      -145.0,  0.0,      1  };
-static constexpr EffectParameter RMSLevel{ &EffectLoudness::mRMSLevel,
-   L"RMSLevel",            -20.0,      -145.0,  0.0,      1  };
-static constexpr EffectParameter DualMono{ &EffectLoudness::mDualMono,
-   L"DualMono",            true,       false,   true,     1  };
-static constexpr EffectParameter NormalizeTo{ &EffectLoudness::mNormalizeTo,
-   L"NormalizeTo",         (int)kLoudness , 0    ,   nAlgos-1, 1  };
-}
 const EffectParameterMethods& EffectLoudness::Parameters() const
 {
    static CapturedParameters<EffectLoudness,
