+++ src/effects/Normalize.cpp
@@ -33,16 +33,6 @@
 #include "../widgets/valnum.h"
 #include "../widgets/ProgressDialog.h"
 
-namespace {
-static constexpr EffectParameter PeakLevel{ &EffectNormalize::mPeakLevel,
-   L"PeakLevel",           -1.0,    -145.0,  0.0,  1  };
-static constexpr EffectParameter RemoveDC{ &EffectNormalize::mDC,
-   L"RemoveDcOffset",      true,    false,   true, 1  };
-static constexpr EffectParameter ApplyGain{ &EffectNormalize::mGain,
-   L"ApplyGain",           true,    false,   true, 1  };
-static constexpr EffectParameter StereoInd{ &EffectNormalize::mStereoInd,
-   L"StereoIndependent",   false,   false,   true, 1  };
-}
 const EffectParameterMethods& EffectNormalize::Parameters() const
 {
    static CapturedParameters<EffectNormalize,
