+++ src/effects/Reverb.cpp
@@ -44,28 +44,6 @@ enum
    ID_WetOnly
 };
 
-namespace {
-static constexpr EffectParameter RoomSize{ &EffectReverb::Params::mRoomSize,
-   L"RoomSize",      75.0,      0,       100,  1  };
-static constexpr EffectParameter PreDelay{ &EffectReverb::Params::mPreDelay,
-   L"Delay",         10.0,      0,       200,  1  };
-static constexpr EffectParameter Reverberance{ &EffectReverb::Params::mReverberance,
-   L"Reverberance",  50.0,      0,       100,  1  };
-static constexpr EffectParameter HfDamping{ &EffectReverb::Params::mHfDamping,
-   L"HfDamping",     50.0,      0,       100,  1  };
-static constexpr EffectParameter ToneLow{ &EffectReverb::Params::mToneLow,
-   L"ToneLow",       100.0,     0,       100,  1  };
-static constexpr EffectParameter ToneHigh{ &EffectReverb::Params::mToneHigh,
-   L"ToneHigh",      100.0,     0,       100,  1  };
-static constexpr EffectParameter WetGain{ &EffectReverb::Params::mWetGain,
-   L"WetGain",       -1.0,      -20,     10,   1  };
-static constexpr EffectParameter DryGain{ &EffectReverb::Params::mDryGain,
-   L"DryGain",       -1.0,      -20,     10,   1  };
-static constexpr EffectParameter StereoWidth{ &EffectReverb::Params::mStereoWidth,
-   L"StereoWidth",   100.0,     0,       100,  1  };
-static constexpr EffectParameter WetOnly{ &EffectReverb::Params::mWetOnly,
-   L"WetOnly",       false,   false,   true, 1  };
-}
 const EffectParameterMethods& EffectReverb::Parameters() const
 {
    static CapturedParameters<EffectReverb,
