+++ src/effects/TimeScale.h
@@ -79,7 +79,6 @@ private:
    void Update_Slider_RatePercentChangeStart();
    void Update_Slider_RatePercentChangeEnd();
 
-public: // TODO remove
    bool bPreview;
    double previewSelectedDuration;
    SlideType slideTypeRate;
@@ -102,6 +101,19 @@ public: // TODO remove
 
    const EffectParameterMethods& Parameters() const override;
    DECLARE_EVENT_TABLE()
+
+static constexpr EffectParameter RatePercentStart{ &EffectTimeScale::m_RatePercentChangeStart,
+   L"RatePercentChangeStart",  0.0,  -90.0,   500,   1  };
+static constexpr EffectParameter RatePercentEnd{ &EffectTimeScale::m_RatePercentChangeEnd,
+   L"RatePercentChangeEnd",    0.0,  -90.0,   500,   1  };
+static constexpr EffectParameter HalfStepsStart{ &EffectTimeScale::m_PitchHalfStepsStart,
+   L"PitchHalfStepsStart",     0.0,  -12.0,   12.0,  1  };
+static constexpr EffectParameter HalfStepsEnd{ &EffectTimeScale::m_PitchHalfStepsEnd,
+   L"PitchHalfStepsEnd",       0.0,  -12.0,   12.0,  1  };
+static constexpr EffectParameter PitchPercentStart{ &EffectTimeScale::m_PitchPercentChangeStart,
+   L"PitchPercentChangeStart", 0.0,  -50.0,   100.0, 1  };
+static constexpr EffectParameter PitchPercentEnd{ &EffectTimeScale::m_PitchPercentChangeEnd,
+   L"PitchPercentChangeEnd",   0.0,  -50.0,   100.0, 1  };
 };
 
 #endif // __AUDACITY_EFFECT_TIMESCALE
