+++ src/effects/ToneGen.h
@@ -15,6 +15,7 @@
 
 #include "Effect.h"
 #include "../ShuttleAutomation.h"
+#include <float.h> // for DBL_MAX
 
 class NumericTextCtrl;
 class ShuttleGui;
@@ -58,7 +59,6 @@ private:
 
    void OnControlUpdate(wxCommandEvent & evt);
 
-public: // TODO remove
    const bool mChirp;
 
    // mSample is an external placeholder to remember the last "buffer"
@@ -82,6 +82,45 @@ public: // TODO remove
 
    const EffectParameterMethods& Parameters() const override;
    DECLARE_EVENT_TABLE()
+
+   enum kWaveforms
+   {
+      kSine,
+      kSquare,
+      kSawtooth,
+      kSquareNoAlias,
+      kTriangle,
+      nWaveforms
+   };
+
+   static const EnumValueSymbol kWaveStrings[nWaveforms];
+
+   enum kInterpolations
+   {
+      kLinear,
+      kLogarithmic,
+      nInterpolations
+   };
+
+   static const EnumValueSymbol kInterStrings[nInterpolations];
+
+// Yes, mFrequency0 and mAmplitude0 are each associated with more than one
+static constexpr EffectParameter StartFreq{ &EffectToneGen::mFrequency0,
+   L"StartFreq",     440.0,   1.0,     DBL_MAX,                1  };
+static constexpr EffectParameter EndFreq{ &EffectToneGen::mFrequency1,
+   L"EndFreq",       1320.0,  1.0,     DBL_MAX,                1  };
+static constexpr EffectParameter StartAmp{ &EffectToneGen::mAmplitude0,
+   L"StartAmp",      0.8,     0.0,     1.0,                    1  };
+static constexpr EffectParameter EndAmp{ &EffectToneGen::mAmplitude1,
+   L"EndAmp",        0.1,     0.0,     1.0,                    1  };
+static constexpr EffectParameter Frequency{ &EffectToneGen::mFrequency0,
+   L"Frequency",     440.0,   1.0,     DBL_MAX,                1  };
+static constexpr EffectParameter Amplitude{ &EffectToneGen::mAmplitude0,
+   L"Amplitude",     0.8,     0.0,     1.0,                    1  };
+static constexpr EnumParameter Waveform{ &EffectToneGen::mWaveform,
+   L"Waveform",      0,       0,       nWaveforms - 1,      1, kWaveStrings, nWaveforms  };
+static constexpr EnumParameter Interp{ &EffectToneGen::mInterpolation,
+   L"Interpolation", 0,       0,       nInterpolations - 1, 1, kInterStrings, nInterpolations  };
 };
 
 class EffectChirp final : public EffectToneGen
