+++ src/effects/TruncSilence.cpp
@@ -72,14 +72,7 @@ using Region = WaveTrack::Region;
 // Declaration of RegionList
 class RegionList : public std::list < Region > {};
 
-enum kActions
-{
-   kTruncate,
-   kCompress,
-   nActions
-};
-
-static const EnumValueSymbol kActionStrings[nActions] =
+const EnumValueSymbol EffectTruncSilence::kActionStrings[nActions] =
 {
    { XO("Truncate Detected Silence") },
    { XO("Compress Excess Silence") }
@@ -93,20 +86,6 @@ static CommandParameters::ObsoleteMap kObsoleteActions[] = {
 
 static const size_t nObsoleteActions = WXSIZEOF( kObsoleteActions );
 
-namespace {
-static constexpr EffectParameter Threshold{ &EffectTruncSilence::mThresholdDB,
-   L"Threshold",  -20.0,      -80.0,   -20.0,                     1  };
-static constexpr EnumParameter ActIndex{ &EffectTruncSilence::mActionIndex,
-   L"Action",     (int)kTruncate,  0,       nActions - 1,           1, kActionStrings, nActions };
-static constexpr EffectParameter Minimum{ &EffectTruncSilence::mInitialAllowedSilence,
-   L"Minimum",    0.5,        0.001,   10000.0,                   1  };
-static constexpr EffectParameter Truncate{ &EffectTruncSilence::mTruncLongestAllowedSilence,
-   L"Truncate",   0.5,        0.0,     10000.0,                   1  };
-static constexpr EffectParameter Compress{ &EffectTruncSilence::mSilenceCompressPercent,
-   L"Compress",   50.0,       0.0,     99.9,                      1  };
-static constexpr EffectParameter Independent{ &EffectTruncSilence::mbIndependent,
-   L"Independent", false,     false,   true,                      1  };
-}
 const EffectParameterMethods& EffectTruncSilence::Parameters() const
 {
    static CapturedParameters<EffectTruncSilence,
